requirements.txt
boto3
itertools
pyyaml
ipaddress
random
